+++ spec/factories/character/items.rb
@@ -2,7 +2,8 @@
 
 FactoryBot.define do
   factory :character_item, class: 'Character::Item' do
-    quantity { 1 }
+    states { Character::Item.default_states }
+    state { 'hands' }
     character
     item
   end